pyproject.toml
[project]
name = "wealth-hub-agent"
version = "0.1.0"
description = "Wealth Hub agent service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  # Core web and runtime deps
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.116.1",
  "mangum>=0.17.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.29.0",
  "pydantic>=2.9.2",
  "pyyaml>=6.0.2",
  "python-jose>=3.5.0",
  # Agent-specific deps (from requirements.txt)
  "openai>=2.14.0",
  "httpx",
  "mcp",
  "finnhub-python>=2.4.26",
  "prompt-toolkit>=3.0.52",
  "python-dotenv>=1.2.1",
  "pyyaml>=6.0.3",
  "rich>=14.2.0",
  "websocket-client>=1.9.0",
  "fastmcp>=2.14.1",
  "langchain-mcp-adapters>=0.2.1",
  "diskcache>=5.6.3",
  "llama-cloud-services>=0.6.88",
  "llama-index-embeddings-openai>=0.5.1",
  "llama-index-llms-openai>=0.6.12",
  "chromadb>=1.4.0",
  "beautifulsoup4>=4.14.3",
  "llama-index-vector-stores-chroma>=0.5.5",
]

[project.optional-dependencies]
dev = [
  "watchfiles>=1.0.0",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "rich>=13.7.0",
]
test = [
  "pytest>=8.4.2",
  "testcontainers>=4.9.0",
  "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM","PL","RUF"]
ignore = ["E501", "PLR0913", "PLC0415", "E402", "PLR0915"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = false
